Moved org.candlepin.subscriptions.db.OrgConfigRepository to org.candlepin.subscriptions.db.OrgConfigRepository
