+++ src/main/java/org/candlepin/subscriptions/retention/TallyRetentionController.java
@@ -23,9 +23,9 @@ package org.candlepin.subscriptions.retention;
 import io.micrometer.core.annotation.Timed;
 import java.time.OffsetDateTime;
 import java.util.stream.Stream;
+import org.candlepin.subscriptions.db.OrgConfigRepository;
 import org.candlepin.subscriptions.db.TallySnapshotRepository;
 import org.candlepin.subscriptions.db.model.Granularity;
-import org.candlepin.subscriptions.db.model.OrgConfigRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
