+++ src/test/java/org/candlepin/subscriptions/security/OptInControllerTest.java
@@ -26,7 +26,7 @@ import static org.mockito.Mockito.*;
 import java.time.OffsetDateTime;
 import java.util.TimeZone;
 import org.candlepin.subscriptions.db.AccountConfigRepository;
-import org.candlepin.subscriptions.db.model.OrgConfigRepository;
+import org.candlepin.subscriptions.db.OrgConfigRepository;
 import org.candlepin.subscriptions.db.model.config.AccountConfig;
 import org.candlepin.subscriptions.db.model.config.OptInType;
 import org.candlepin.subscriptions.db.model.config.OrgConfig;
