+++ src/test/java/org/candlepin/subscriptions/subscription/SubscriptionPruneControllerTest.java
@@ -31,9 +31,9 @@ import io.micrometer.core.instrument.Timer;
 import java.util.HashMap;
 import java.util.stream.Stream;
 import org.candlepin.subscriptions.capacity.files.ProductDenylist;
+import org.candlepin.subscriptions.db.OrgConfigRepository;
 import org.candlepin.subscriptions.db.SubscriptionRepository;
 import org.candlepin.subscriptions.db.model.Offering;
-import org.candlepin.subscriptions.db.model.OrgConfigRepository;
 import org.candlepin.subscriptions.db.model.Subscription;
 import org.candlepin.subscriptions.db.model.SubscriptionMeasurementKey;
 import org.candlepin.subscriptions.task.TaskQueueProperties;
